--- Tests/JWTKitTests/Utils/Bundle+Sendable.swift
@@ -1,5 +1,0 @@
-import Foundation
-
-#if compiler(<6.0) && !canImport(Darwin)
-extension Foundation.Bundle: @unchecked Swift.Sendable {}
-#endif
