--- Tests/JWTKitTests/Utils/String+bytes.swift
@@ -1,5 +1,0 @@
-extension String {
-    var bytes: [UInt8] {
-        .init(utf8)
-    }
-}
