--- Tests/JWTKitTests/Utils/TestPayload.swift
@@ -1,12 +1,0 @@
-import JWTKit
-
-struct TestPayload: JWTPayload, Equatable {
-    var sub: SubjectClaim
-    var name: String
-    var admin: Bool
-    var exp: ExpirationClaim
-
-    func verify(using _: some JWTAlgorithm) throws {
-        try exp.verifyNotExpired()
-    }
-}
